+++ src/hooks/useMFA.tsx
@@ -38,17 +38,17 @@ export function useMFA() {
     if (hasMFA) {
       const factors = await supabase.auth.mfa.listFactors();
 
-      sessionStorage.setItem('@uollet:hasMFA', 'true');
-
       if (factors.error) {
-        toast.error(factors.error.message);
+        toast.error('Erro ao verificar fatores MFA. Tente novamente', {
+          id: 'error',
+        });
         return;
       }
 
       const totpFactor = factors.data.totp[0];
 
       if (!totpFactor) {
-        toast.error('Nenhum fator TOTP encontrado!');
+        toast.error('Nenhum fator TOTP encontrado!', { id: 'error' });
         return;
       }
 
@@ -70,15 +70,23 @@ export function useMFA() {
       });
 
       if (verify.error) {
-        toast.error(verify.error.message);
+        toast.error(`Erro ao verificar código: ${verify.error.message}`, {
+          id: 'error',
+        });
         return;
       }
 
-      toast.success('Bem vindo ao uollet', {
-        id: 'success',
-      });
+      if (!verify.data) return;
 
-      navigate('/');
+      if (verify.data) {
+        sessionStorage.setItem('@uollet:hasMFA', 'true');
+
+        toast.success('Bem vindo ao uollet', {
+          id: 'success',
+        });
+
+        navigate('/');
+      }
     }
   }