+++ src/utils/category.ts
@@ -41,9 +41,10 @@ import {
   nomad,
   revolut,
   will,
+  money,
+  savings,
+  goals,
 } from 'assets';
-import money from 'assets/categories/money.svg';
-import piggyBank from 'assets/categories/piggy-bank.svg';
 
 export const category = [
   {
@@ -272,8 +273,13 @@ export const category = [
     category: 'bank',
   },
   {
-    icon: piggyBank,
+    icon: savings,
     name: 'Rendimentos',
     category: 'savings',
+  },
+  {
+    icon: goals,
+    name: 'Meta',
+    category: 'goals',
   },
 ];